lightnvm: manage block list on LUN owner
LUNs can be exclusively owned by a target through the media manager's
reserve_lun function. In this case, the target should implement its own
provisioning and manage internally the free/used/bad block list.

This patch introduces a LUN management structure that can be passed on
to LUN exclusive owners. The media manager is the default owner. On boot
up, it populates the lists with based on the device's bad block table
and sysblk metadata. Then, if the LUN is owned by a target, the
management structured is passed on to it. From this moment, the target
is responsible for list management.

This simplifies target-specific block management logic. For example, GC
is strictly handled by the target. The media manager should therefore
not reserve blocks for GC or any other purpose.

Adapt rrpc to the new reserve_lun/release_lun interface and move the
logic for reserving blocks for GC to the target.

Signed-off-by: Javier González <javier@cnexlabs.com>